+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `swap_tiles` to `RegularBackground` and `AffineBackground` for cheaply moving around tiles.
+
 ## [0.24.0] - 2026/06/10
 
 ### Added